Renaming AssertUtils->ObjectTransactionAssertions
Since it is POT specific. It is not a general utility
