+++ arena-pot/src/main/scala/com/uqbar/common/transaction/AssertUtils.java
@@ -1,28 +1,10 @@
 package com.uqbar.common.transaction;
 
-
 /**
  * 
  * @author jfernandes
+ * @deprecated use {@link ObjectTransactionAssertions}
  */
-public class AssertUtils {
-
-	public static void assertNull(String message, Object obj) {
-		assertTrue(message, obj == null);
-	}
-
-	public static void assertNotNull(String message, Object obj) {
-		assertTrue(message, obj != null);
-	}
-
-	public static void assertTrue(String message, boolean condition) {
-		if (!condition) {
-			throw new ObjectTransactionException(message);
-		}
-	}
-
-	public static void assertArgumentNotNull(String message, Object argument) {
-		assertNotNull(message, argument);
-	}
-
+@Deprecated
+public class AssertUtils extends ObjectTransactionAssertions {
 }
